requirements.txt
pandas
numpy
scikit-learn
nltk
sentence-transformers
faiss-cpu
tqdm
